Removed the drop( i.vars ) from estimates table call
This command can be problematic, for instance, when there are no observations, and also when partial( i.vars) is used in ivreg2.
The simplest thing is to just remove this feature, which is meant to avoid a really long list of coefficients for fixed effects,
and somewhere else have code that avoids rendering those values in tables.